EmgGest/Server/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
matplotlib==3.10.0
numpy==1.24.2
pandas==2.2.3
Pillow==11.1.0
scipy==1.15.0
torch==2.3.0
torchvision==0.18.0
